backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
pymongo==4.6.1
opencv-python==4.8.1.78
insightface==0.7.3
